Refactor source by feature at top level
- Removed explicit references to index in imports
