+++ penny-guess-client/src/main/HighScores/components/HighScoreTable.jsx
