+++ penny-guess-client/src/main/HighScores/components/HighScores.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import api from "../../api";
+import api from "../../api/index";
 import HighScoreTable from './HighScoreTable';
 
 export default class HighScores extends React.Component {
