+++ penny-guess-client/src/main/HighScores/index.js
@@ -1,0 +1,1 @@
+export { default } from './components/HighScores';
