+++ penny-guess-client/src/main/Home/index.js
@@ -1,0 +1,1 @@
+export { default } from './components/Home';
