+++ penny-guess-client/src/main/Snake/components/SnakeGame.jsx
