+++ penny-guess-client/src/main/Snake/components/SnakeSubmitHighScore.jsx
