+++ penny-guess-client/src/main/Snake/index.js
@@ -1,0 +1,1 @@
+export { default } from './components/Snake';
