+++ penny-guess-client/src/main/Snake/schemes.js
