--- penny-guess-client/src/main/components/HighScores/index.js
@@ -1,1 +1,0 @@
-export { default } from './HighScores';
