--- penny-guess-client/src/main/components/Snake/index.js
@@ -1,1 +1,0 @@
-export { default } from './Snake';
