--- penny-guess-client/src/main/components/index.js
@@ -1,3 +1,0 @@
-export { default as Home } from './Home/index';
-export { default as Snake } from './Snake/index';
-export { default as HighScores } from './HighScores/index';
