+++ penny-guess-client/src/main/index.js
@@ -1,0 +1,2 @@
+export { default as Home } from './Home';
+export { default as HighScores } from './HighScores';
